requirements.txt
# Function dependencies, for example:
# package>=version
elasticsearch==7.10.0
elasticsearch-dsl==7.4.0
wheel==0.38.1
python-dateutil==2.8.2
urllib3==1.26.14
python-dotenv
tldextract